chore 📦: Update package versions and add new dependencies
This commit updates several packages to their latest versions, including starlette, typer, treetable, and types-pyyaml. Additionally, it adds the 'typing-extensions' package with version 4.15.0 and includes a new dependency 'types-pyyaml' with version 6.0.12.20250915. These changes ensure that the project remains up-to-date with the latest features and bug fixes, while also introducing new functionalities and dependencies to enhance the project's capabilities.