chore 📦: Update project version to 1.0.1 and add dependencies
This commit updates the project version from 1.0.0 to 1.0.1 in the `pyproject.toml` file. Additionally, it adds the `torchaudio==2.0.1` dependency to the `bench-torch-2-0-1` benchmark configuration and the `torchaudio==2.3.0`, `torch==2.3.0`, `torchaudio==2.4.1`, and `torch==2.4.1` dependencies to the `bench-torch-2-3-0`, `bench-torch-2-4-1`, and `bench-torch-2-6-0` benchmarks respectively. These changes ensure that the project can run the specified benchmarks with the required versions of the libraries, enhancing the reliability and performance of the tests.